+++ ssha/ssh.py
@@ -10,21 +10,15 @@ from . import config
 
 def _get_address(instance_ip):
 
-    name = config.get('ssh.user.name')
-    prefix = config.get('ssh.user.prefix')
+    username = config.get('ssh.username_prefix', '') + config.get('ssh.username', '')
 
     # Don't add the username to the address when it is the current user,
     # because it would make no difference.
-    user = os.environ.get('USER')
-    if name == user:
-        name = None
+    if username == os.environ.get('USER'):
+        username = None
 
-    if not name and prefix:
-        return prefix + user + '@' + instance_ip
-    elif name and prefix:
-        return prefix + name + '@' + instance_ip
-    elif name:
-        return name + '@' + instance_ip
+    if username:
+        return username + '@' + instance_ip
     else:
         return instance_ip
